3-langgraph/Agentic-workflow-project/requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.1

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
unstructured>=0.10.0
python-pptx>=0.6.21

# Vector stores and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Web search and tools
tavily-python>=0.3.0

# UI and visualization
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional dependencies for better functionality
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0